refactor mission file content and ensure mission test passes
